Muevo clase Spotify y TokenCache a modulo 'recursos'
